requirements.txt
click==8.1.3
Flask==2.2.2
Flask-Login==0.6.2
Flask-SQLAlchemy==2.5.1
greenlet==1.1.3
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
python-dotenv==0.21.0
SQLAlchemy==1.4.41
Werkzeug==2.2.2
